Consolidate chat schema and validation in chat_normalize
Move the chat schema and validators into chat_normalize and DatakitChatSource into sft_sources. Remove chat.py and update all imports, keeping normalization behavior unchanged.
