+++ lib/marin/src/marin/datakit/download/swe_rebench_openhands.py
@@ -16,8 +16,7 @@ from zephyr import counters
 from zephyr.context import ZephyrContext
 from zephyr.dataset import Dataset
 
-from marin.datakit.chat import CHAT_SCHEMA
-from marin.datakit.chat_normalize import normalize_chat_step
+from marin.datakit.chat_normalize import CHAT_SCHEMA, normalize_chat_step
 from marin.datakit.download.huggingface import download_hf_step
 from marin.datakit.download.opencode import INLINE_TOOL_CALL
 from marin.datakit.download.rollout_transforms import (
