+++ lib/marin/src/marin/datakit/download/synthetic1.py
@@ -14,8 +14,7 @@ from zephyr.context import ZephyrContext
 from zephyr.dataset import Dataset
 from zephyr.readers import load_parquet
 
-from marin.datakit.chat import CHAT_SCHEMA
-from marin.datakit.chat_normalize import normalize_chat_step
+from marin.datakit.chat_normalize import CHAT_SCHEMA, normalize_chat_step
 from marin.datakit.download.huggingface import download_hf_step
 from marin.datakit.download.rollout_transforms import checked_openai_chat_document, strip_think_tags, text_document
 from marin.datakit.normalize import normalize_step
